developing java client
git-svn-id: file:///home/amok/repo/galaxy/trunk@2213 e5d723ff-ac09-4e1e-99b2-b35108d02165

+++ java_client/Xpm.java
@@ -590,7 +590,7 @@ public class Xpm {
 				}
 				pixchar = new Integer( tmpchar & 0xffff );
 
-				pixval = ( Integer )colorlookup.get( pixchar );
+				pixval = colorlookup.get( pixchar );
 				if ( pixval == null ) {
 					if ( debugflag ) {
 						debug( "HEY MORON: no value stored "+
